+++ packages/moonstone/Panels/index.js
@@ -19,6 +19,7 @@ import ActivityPanels from './ActivityPanels';
 import AlwaysViewingPanels from './AlwaysViewingPanels';
 import Breadcrumb from './Breadcrumb';
 import Header, {HeaderBase} from './Header';
+import MinimizeController, {MinimizeControllerContext} from './MinimizeController';
 import Panel from './Panel';
 import Panels from './Panels';
 
@@ -29,6 +30,8 @@ export {
 	Breadcrumb,
 	Header,
 	HeaderBase,
+	MinimizeController,
+	MinimizeControllerContext,
 	Panel,
 	Panels,
 	Panels as PanelsBase,